+++ src/Query/Criteria/Filter/Range.php
@@ -3,14 +3,14 @@ declare(strict_types=1);
 
 namespace Kununu\Elasticsearch\Query\Criteria\Filter;
 
-class Range
+final class Range
 {
     public const string KEYWORD = 'range';
 
     public static function asArray(string $field, array $value, array $options = []): array
     {
         return [
-            static::KEYWORD => [
+            self::KEYWORD => [
                 $field => array_merge(
                     $options,
                     $value