+++ src/Query/Criteria/Filter/Regexp.php
@@ -3,14 +3,14 @@ declare(strict_types=1);
 
 namespace Kununu\Elasticsearch\Query\Criteria\Filter;
 
-class Regexp
+final class Regexp
 {
     public const string KEYWORD = 'regexp';
 
     public static function asArray(string $field, $value, array $options = []): array
     {
         return empty($options)
-            ? [static::KEYWORD => [$field => $value]]
-            : [static::KEYWORD => [$field => array_merge($options, ['value' => $value])]];
+            ? [self::KEYWORD => [$field => $value]]
+            : [self::KEYWORD => [$field => array_merge($options, ['value' => $value])]];
     }
 }